[#12] fix: Change x,y type from Long to Double
경도, 위도 값을 실수로 Long으로 선언하여 소수점을 못받고 있었습니다.

+++ BE/airbnb/src/main/java/team14/airbnb/domain/dto/request/accommodation/AccommodationAddressDto.java
@@ -27,10 +27,10 @@ public class AccommodationAddressDto {
     private String region3DepthName;
 
     @NotNull(message = "주소의 경도값은 필수입니다.")
-    private Long x;
+    private Double x;
 
     @NotNull(message = "주소의 위도값은 필수입니다.")
-    private Long y;
+    private Double y;
 
     public AccommodationAddress toEntity() {
         return new AccommodationAddress(
